Handle Paypal missing payment status types.
The following payment status types are being added:
- Canceled_Reversal
- Refunded
- Reversed
- Voided

The payment_status field in PayPalStandardBase is being
modified. The max_length is now 17 to handle new types.
Corresponding migrations for ipn and pdt exist.

The following new IPN signals are being added:
- payment_was_refunded
- payment_was_reversed

The following new methods are being added to PayPalStandardBase:
- is_refund
- is_reversed

Added test for the new signals.